sound: extract init and shutdown to trx
